# docs/openapi/shopflow-aquatrade-openapi.yaml
openapi: 3.0.3
info:
  title: ShopFlow AquaTrade Integration API
  version: 1.0.0
servers:
  - url: http://localhost
paths:
  /api/orders:
    get:
      summary: List tenant orders
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
  /api/orders/{orderId}:
    get:
      summary: Order detail with BOLA protection
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Cross-tenant blocked }
  /api/catalog/lots:
    get:
      summary: List AquaTrade lots by tenant
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/vendors/me:
    get:
      summary: Current tenant vendor profile
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/shipments:
    get:
      summary: List tenant shipments
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/quotes:
    post:
      summary: Create quote for a lot
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lot_id, quantity_kg]
              properties:
                lot_id: { type: string }
                quantity_kg: { type: number }
      responses:
        '201': { description: Created }
        '403': { description: Cross-tenant blocked }
  /api/auth/refresh:
    post:
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200': { description: Refreshed }
        '401': { description: Replay or invalid refresh token }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT